+++ packages/motion/template/main.js
@@ -1,8 +1,6 @@
 /* @flow */
 
-import Motion from 'react'
-
-class Main extends Motion.Component {
+class Main extends React.Component {
   render() {
     return <div>Hello World</div>
   }