+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Upcoming
 
-- Simplify babel-preset by importing `babel-preset-react`
-- Remove `$` support in import, in favor of auto-replacement in atom package
+- Simplify babel-preset by importing `babel-preset-react` (**API BREAKING**)
+- Remove `$` support in import, in favor of auto-replacement in atom package (**API BREAKING**)
 
 ## 1.4.1